Initial Commit
Setting things up. Created html files for all the site pages. Created the navigation bar on the homepage. Started work on a centering container. Created and added the animated slideshow script and mobile icon script to the homepage. Created a stylesheet with navbar, icon, slideshow, animations, and responsive mobile styling.