Homepage
Finished the homepage as per Nan's draft. Added a content container for margins and a set height. Added a centered responsive image and  descriptive text. Padded the new elements for desktop and mobile. Edited the responsive nav icon so it always stays in the corner, instead of moving when colliding with the title. Changed some header text colors. Changed the footer to relative positioning.