About page + Fixes
Created the About page with Nan's image and descriptor text. Fixed the navigation script not functioning due to renamed classes. Added the content container to the tours page. Added secondary page classes to make page specific styling changes. Lots of styling changes. Made the site images 10% bigger. Made the content container slightly bigger to take into account the side padding. Added a top margin to the content container for it to take care of content spacing, instead of whatever is the first element on each page, which was janky. Edited general header 2 spacing, and added page specific spacing for header 2s and images on the about and videos pages. Made the header title inline block so it moves out of the way of the nav icon without overlapping. Changed some mobile padding and margins to be more compact and consistant across the site.