Emal Updates
Added a new Portfolio page to the site, and added it to the navigation bars on each page. Added the new image and revised text to the about page, with a new container so the image and caption text stay together.  Moved the new image on the lectures page to the top and added a few more lines of text. Added a new image on the publications page and added the revised text and links. Added a contact sentence to the tours page. Brought the page content closer to the navigation bar, as I felt there was too much empty space that site visitors would have to scroll down through. Added a class for image caption text that is smaller, lighter, and italicized. Styled the new about image container for desktop and mobile.