Finished Portfolio and Press
Added all the images and caption text to the portfolio and press grids. Added a class to the grid images so they don't get overwritten by the content container image styling. Edited padding for the grid and grid text, and made the font slightly smaller. Moved the about page to the index file and edited the navigation links on all the pages.