Start SEO Optimization
Started some SEO optimization for the site. Added nofollow to all the off site links, and made them open in a new tab.  Added some keywords to start with on the index page. Added a robots.txt file, needs a correct sitemap. Added a contact section to the about page, and a link to Nan's resume. Styled the header and link for the about page.