Email Updates
Completed about half of the email edits and updates Nan sent. Deleted the research page and added the about page back in. Changed the site's font to the one Nan sent. Updated and reordered the navigation links. Changed the title and navigation colors, capitilized and underlined the navigation text, and greatly decreased the spacing and size of the whole header navigation. Made the homepage a simple image, and moved all its information back to the about page. Made Nan's edits to the text on the about page. Added a commented colors reference at the top of the style sheet. Next up is making a links section on the about page, either as text or as font awesome icons.