More Email Updates
Small site restructuring. Removed the videos and press pages from the navigation bars, and added a new blank Contact page. Moved all the links on the about page to the bottom,  added new text, and deleted old text. Added a twitter icon and link. Changed the image to be centered above the text and show three new inline images, with new styling. Updated the CV. Fixed some typos on lectures and tours. Redid the Print publications section and removed the image links and descriptive text, and changed it to a simple links section. Moved the videos to the tours page. Darkened most font colors by 10%-25%. Made the image grid only expand to two columns instead of three. Edited some image sizes on mobile. Still need to change the font to Cablibri.