Finish Updates + Meta Tags
Removed the about, press, and video pages from the site and put them in a storage folder. Moved the about page back to the home page, and changed all the navigation links to match the new site structure. Changed the main font for the site to Calibri. Added alt text to all the images. Created a placeholder contact page until Nan tells me what she want on it. Filled out all the necessary meta tags on the index file. Added a simple NT favicon for the site and the appropriate meta information. Made the publication image into the Facebook preview favicon. Filled out keywords for each page. Wrote the description for the index, need to finish the tags for the rest of the pages still.