Final Email Updates
Should be the last email updates push. Reverted the changes combining the about page and homepage. Readded the about page to the navigation bar, and switched the order of tours and portfolio. Corrected a typo on the about page, and removed the contact and links sections. Added a right-aligned image to the contact page. Added a new paragraph to the publications page, and changed the image caption text. Added favicon, description, and url meta information to all the the pages that didn't have it. Added a sitemap, though it might need to be changed once the site is uploaded online.